test: Lift test store helpers to solver package
We will use these helpers for top-level solver integration tests in
addition to the store integration tests.
