+++ pkg/solver/store/store_test.go
@@ -11,23 +11,20 @@ import (
 	"time"
 
 	"github.com/lilypad-tech/lilypad/pkg/data"
+	"github.com/lilypad-tech/lilypad/pkg/solver"
 	"github.com/lilypad-tech/lilypad/pkg/solver/store"
 	solverstore "github.com/lilypad-tech/lilypad/pkg/solver/store"
-	databasestore "github.com/lilypad-tech/lilypad/pkg/solver/store/db"
-	memorystore "github.com/lilypad-tech/lilypad/pkg/solver/store/memory"
 	"golang.org/x/exp/rand"
 )
-
-const DB_CONN_STR = "postgres://postgres:postgres@localhost:5432/solver-db?sslmode=disable"
 
 // Job offers
 
 func TestJobOfferOps(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test multiple job offers in a single test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -221,14 +218,14 @@ func TestJobOfferQuery(t *testing.T) {
 		},
 	}
 
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
-		getStore, clearStore := config.init()
+		getStore, clearStore := config.Init()
 		defer clearStore()
 
 		for _, tc := range testCases {
 			// Test each case in a separate test run
-			t.Run(fmt.Sprintf("%s/%s", config.name, tc.name), func(t *testing.T) {
+			t.Run(fmt.Sprintf("%s/%s", config.Name, tc.name), func(t *testing.T) {
 				store := getStore()
 
 				// Add test job offers
@@ -267,12 +264,12 @@ func TestJobOfferSort(t *testing.T) {
 	const delay = 50 * time.Millisecond
 	numOffers := 3 + rand.Intn(5) // Random offers between 3 and 7
 
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
-		getStore, clearStore := config.init()
+		getStore, clearStore := config.Init()
 		defer clearStore()
 
-		t.Run(config.name, func(t *testing.T) {
+		t.Run(config.Name, func(t *testing.T) {
 			store := getStore()
 
 			// Track IDs in insertion order
@@ -319,11 +316,11 @@ func TestJobOfferSort(t *testing.T) {
 // Resource Offer
 
 func TestResourceOfferOps(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test multiple resource offers in a single test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -529,14 +526,14 @@ func TestResourceOfferQuery(t *testing.T) {
 		},
 	}
 
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
-		getStore, clearStore := config.init()
+		getStore, clearStore := config.Init()
 		defer clearStore()
 
 		for _, tc := range testCases {
 			// Test each case in a separate test run
-			t.Run(fmt.Sprintf("%s/%s", config.name, tc.name), func(t *testing.T) {
+			t.Run(fmt.Sprintf("%s/%s", config.Name, tc.name), func(t *testing.T) {
 				store := getStore()
 
 				// Add test resource offers
@@ -575,12 +572,12 @@ func TestResourceOfferSort(t *testing.T) {
 	const delay = 50 * time.Millisecond
 	numOffers := 3 + rand.Intn(5) // Random offers between 3 and 7
 
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
-		getStore, clearStore := config.init()
+		getStore, clearStore := config.Init()
 		defer clearStore()
 
-		t.Run(config.name, func(t *testing.T) {
+		t.Run(config.Name, func(t *testing.T) {
 			store := getStore()
 
 			// Track IDs in insertion order
@@ -627,11 +624,11 @@ func TestResourceOfferSort(t *testing.T) {
 // Deals
 
 func TestDealOps(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test multiple deals in a single test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -680,11 +677,11 @@ func TestDealOps(t *testing.T) {
 }
 
 func TestDealGetAll(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test batch of deals in a test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -731,11 +728,11 @@ func TestDealGetAll(t *testing.T) {
 }
 
 func TestDealUpdates(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test multiple deals in a single test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -935,14 +932,14 @@ func TestDealQuery(t *testing.T) {
 		},
 	}
 
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
-		getStore, clearStore := config.init()
+		getStore, clearStore := config.Init()
 		defer clearStore()
 
 		for _, tc := range testCases {
 			// Test each case in a separate test run
-			t.Run(fmt.Sprintf("%s/%s", config.name, tc.name), func(t *testing.T) {
+			t.Run(fmt.Sprintf("%s/%s", config.Name, tc.name), func(t *testing.T) {
 				store := getStore()
 
 				// Add deals
@@ -980,11 +977,11 @@ func TestDealQuery(t *testing.T) {
 // Results
 
 func TestResultOps(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test multiple results in a single test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -1079,11 +1076,11 @@ func TestResultOps(t *testing.T) {
 // Match decisions
 
 func TestMatchDecisionOps(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test multiple match decisions in a single test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -1202,10 +1199,10 @@ func TestMatchDecisionOps(t *testing.T) {
 }
 
 func TestMatchDecisionRemove(t *testing.T) {
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -1269,11 +1266,11 @@ func TestConcurrentOps(t *testing.T) {
 	results := generateResults(4, 10)
 	matchDecisions := generateMatchDecisions(4, 10)
 
-	storeConfigs := setupStores(t)
+	storeConfigs := solver.SetupTestStores(t)
 	for _, config := range storeConfigs {
 		// Test concurrent adds in a single test run
-		t.Run(config.name, func(t *testing.T) {
-			getStore, clearStore := config.init()
+		t.Run(config.Name, func(t *testing.T) {
+			getStore, clearStore := config.Init()
 			store := getStore()
 			defer clearStore()
 
@@ -1444,124 +1441,6 @@ func TestConcurrentOps(t *testing.T) {
 			}
 
 		})
-	}
-}
-
-// Utilities
-
-type storeConfig struct {
-	name string
-	init func() (getStore func() store.SolverStore, clearStore func())
-}
-
-func setupStores(t *testing.T) []storeConfig {
-	initMemory := func() (func() store.SolverStore, func()) {
-		// Get store function creates a new memory store
-		// which effectively clears data between runs
-		getStore := func() store.SolverStore {
-			s, err := memorystore.NewSolverStoreMemory()
-			if err != nil {
-				t.Fatalf("Failed to create memory store: %v", err)
-			}
-			return s
-		}
-		clearStore := func() {}
-
-		return getStore, clearStore
-	}
-
-	initDatabase := func() (func() store.SolverStore, func()) {
-		db, err := databasestore.NewSolverStoreDatabase(DB_CONN_STR, "silent")
-		if err != nil {
-			t.Fatalf("Failed to create database store: %v", err)
-		}
-
-		// Clear data at initialization
-		clearStoreDatabase(t, db)
-
-		// Get store functions clear data and returns
-		// the same store instance
-		getStore := func() store.SolverStore {
-			clearStoreDatabase(t, db)
-			return db
-		}
-		clearStore := func() { clearStoreDatabase(t, db) }
-
-		return getStore, clearStore
-	}
-
-	return []storeConfig{
-		{name: "memory", init: initMemory},
-		{name: "database", init: initDatabase},
-	}
-}
-
-func clearStoreDatabase(t *testing.T, s store.SolverStore) {
-	// Delete job offers
-	jobOffers, err := s.GetJobOffers(store.GetJobOffersQuery{
-		IncludeCancelled: true,
-	})
-	if err != nil {
-		t.Fatalf("Failed to get existing job offers: %v", err)
-	}
-
-	for _, result := range jobOffers {
-		err := s.RemoveJobOffer(result.ID)
-		if err != nil {
-			t.Fatalf("Failed to remove existing job offer: %v", err)
-		}
-	}
-
-	// Delete resource offers
-	resourceOffers, err := s.GetResourceOffers(store.GetResourceOffersQuery{})
-	if err != nil {
-		t.Fatalf("Failed to get existing resource offers: %v", err)
-	}
-
-	for _, result := range resourceOffers {
-		err := s.RemoveResourceOffer(result.ID)
-		if err != nil {
-			t.Fatalf("Failed to remove existing resource offer: %v", err)
-		}
-	}
-
-	// Delete deals
-	deals, err := s.GetDealsAll()
-	if err != nil {
-		t.Fatalf("Failed to get existing deals: %v", err)
-	}
-
-	for _, deal := range deals {
-		err := s.RemoveDeal(deal.ID)
-		if err != nil {
-			t.Fatalf("Failed to remove existing deal: %v", err)
-		}
-	}
-
-	// Delete results
-	results, err := s.GetResults()
-	if err != nil {
-		t.Fatalf("Failed to get existing results: %v", err)
-	}
-
-	for _, result := range results {
-		err := s.RemoveResult(result.DealID)
-		if err != nil {
-			t.Fatalf("Failed to remove existing result: %v", err)
-		}
-	}
-
-	// Delete match decisions
-	decisions, err := s.GetMatchDecisions()
-	if err != nil {
-		t.Fatalf("Failed to get existing match decisions: %v", err)
-	}
-
-	for _, decision := range decisions {
-		err := s.RemoveMatchDecision(decision.ResourceOffer, decision.JobOffer)
-		if err != nil {
-			t.Fatalf("Failed to remove existing match decision: %v", err)
-		}
 	}
 }
